Add trash database

+++ includes/functions.inc.php
@@ -134,6 +134,17 @@ function loginUser($conn, $username, $pwd) {
 }
 
 function deleteRecord($conn, $id) {
+    $sql = "INSERT INTO trash SELECT * FROM bugs WHERE id = ?;";
+    $stmt = mysqli_stmt_init($conn);
+    if (!mysqli_stmt_prepare($stmt, $sql)) {
+        header("location: ../add.php?error=stmtfailed");
+        exit();
+    }
+
+    mysqli_stmt_bind_param($stmt, "i", $id);
+    mysqli_stmt_execute($stmt);
+    mysqli_stmt_close($stmt);
+
     $sql = "DELETE FROM bugs WHERE id = ?;";
     $stmt = mysqli_stmt_init($conn);
     if (!mysqli_stmt_prepare($stmt, $sql)) {
